4.21

+++ controllers/blogs.js
@@ -36,8 +36,20 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (req, res, next) => {
   try {
-    await Blog.findByIdAndRemove(req.params.id)
-    res.status(204).end()
+    const decodedToken = jwt.verify(req.token, process.env.SECRET)
+    if (!decodedToken.id) {
+      return res.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const blog = await Blog.findById(req.params.id)
+
+    if (blog.user.toString() === decodedToken.id) {
+      await Blog.findByIdAndRemove(req.params.id)
+      res.status(204).end()
+    } else {
+      return res.status(401).json({ error: 'blog must be deleted by owner' })
+    }
+
   } catch (exception) {
     next(exception)
   }
